FAB Logic working
