+++ lib/post/widgets/comment_view.dart
@@ -49,7 +49,8 @@ class CommentSubview extends StatefulWidget {
   State<CommentSubview> createState() => _CommentSubviewState();
 }
 
-class _CommentSubviewState extends State<CommentSubview> with SingleTickerProviderStateMixin {
+class _CommentSubviewState extends State<CommentSubview>
+    with SingleTickerProviderStateMixin {
   Set collapsedCommentSet = {}; // Retains the collapsed state of any comments
   bool _animatingOut = false;
   bool _animatingIn = false;
@@ -74,7 +75,8 @@ class _CommentSubviewState extends State<CommentSubview> with SingleTickerProvid
       if (status == AnimationStatus.completed && _animatingOut) {
         _animatingOut = false;
         _removeViewFullCommentsButton = true;
-        context.read<PostBloc>().add(const GetPostCommentsEvent(commentParentId: null, viewAllCommentsRefresh: true));
+        context.read<PostBloc>().add(const GetPostCommentsEvent(
+            commentParentId: null, viewAllCommentsRefresh: true));
       }
     });
   }
@@ -90,94 +92,123 @@ class _CommentSubviewState extends State<CommentSubview> with SingleTickerProvid
     }
 
     return ListView.builder(
-        addSemanticIndexes: false,
-        controller: widget.scrollController,
-        itemCount: getCommentsListLength(),
-        itemBuilder: (context, index) {
-          if (widget.postViewMedia != null && index == 0) {
-            return PostSubview(selectedCommentId: widget.selectedCommentId, useDisplayNames: state.useDisplayNames, postViewMedia: widget.postViewMedia!);
-          }
-          if (widget.hasReachedCommentEnd == false && widget.comments.isEmpty) {
-            return Column(
+      addSemanticIndexes: false,
+      controller: widget.scrollController,
+      itemCount: getCommentsListLength(),
+      itemBuilder: (context, index) {
+        if (widget.postViewMedia != null && index == 0) {
+          return PostSubview(
+              selectedCommentId: widget.selectedCommentId,
+              useDisplayNames: state.useDisplayNames,
+              postViewMedia: widget.postViewMedia!);
+        }
+        if (widget.hasReachedCommentEnd == false && widget.comments.isEmpty) {
+          return Column(
+            children: [
+              Container(
+                padding: const EdgeInsets.symmetric(vertical: 24.0),
+                child: const CircularProgressIndicator(),
+              ),
+            ],
+          );
+        } else {
+          return SlideTransition(
+            position: _fullCommentsOffsetAnimation,
+            child: Column(
               children: [
-                Container(
-                  padding: const EdgeInsets.symmetric(vertical: 24.0),
-                  child: const CircularProgressIndicator(),
-                ),
-              ],
-            );
-          } else {
-            return SlideTransition(
-                position: _fullCommentsOffsetAnimation,
-                child: Column(children: [
-                  if (widget.selectedCommentId != null && !_animatingIn && index != widget.comments.length + 1)
-                    Center(
-                        child: Column(children: [
-                      Row(children: [
-                        const Padding(padding: EdgeInsets.only(left: 15)),
-                        Expanded(
-                            child: ElevatedButton(
-                          style: ElevatedButton.styleFrom(
-                            minimumSize: const Size.fromHeight(50),
-                            backgroundColor: theme.colorScheme.primaryContainer,
-                            textStyle: theme.textTheme.titleMedium?.copyWith(
-                              color: theme.colorScheme.primary,
+                if (widget.selectedCommentId != null &&
+                    !_animatingIn &&
+                    index != widget.comments.length + 1)
+                  Center(
+                    child: Column(
+                      children: [
+                        Row(
+                          children: [
+                            const Padding(padding: EdgeInsets.only(left: 15)),
+                            Expanded(
+                              child: ElevatedButton(
+                                style: ElevatedButton.styleFrom(
+                                  minimumSize: const Size.fromHeight(50),
+                                  backgroundColor:
+                                      theme.colorScheme.primaryContainer,
+                                  textStyle:
+                                      theme.textTheme.titleMedium?.copyWith(
+                                    color: theme.colorScheme.primary,
+                                  ),
+                                ),
+                                onPressed: () {
+                                  _animatingOut = true;
+                                  _fullCommentsAnimation.forward();
+                                },
+                                child: const Text('View all comments'),
+                              ),
                             ),
-                          ),
-                          onPressed: () {
-                            _animatingOut = true;
-                            _fullCommentsAnimation.forward();
-                          },
-                          child: const Text('View all comments'),
-                        )),
-                        const Padding(padding: EdgeInsets.only(right: 15))
-                      ]),
-                      const Padding(padding: EdgeInsets.only(top: 10)),
-                    ])),
-                  if (index != widget.comments.length + 1)
-                    CommentCard(
-                      now: widget.now,
-                      selectCommentId: widget.selectedCommentId,
-                      selectedCommentPath: widget.selectedCommentPath,
-                      commentViewTree: widget.comments[index - 1],
-                      collapsedCommentSet: collapsedCommentSet,
-                      collapsed: collapsedCommentSet.contains(widget.comments[index - 1].commentView!.comment.id) || widget.level == 2,
-                      onSaveAction: (int commentId, bool save) => widget.onSaveAction(commentId, save),
-                      onVoteAction: (int commentId, VoteType voteType) => widget.onVoteAction(commentId, voteType),
-                      onCollapseCommentChange: (int commentId, bool collapsed) => onCollapseCommentChange(commentId, collapsed),
-                      onDeleteAction: (int commentId, bool deleted) => widget.onDeleteAction(commentId, deleted),
+                            const Padding(padding: EdgeInsets.only(right: 15))
+                          ],
+                        ),
+                        const Padding(padding: EdgeInsets.only(top: 10)),
+                      ],
                     ),
-                  if (index == widget.comments.length + 1) ...[
-                    if (widget.hasReachedCommentEnd == true) ...[
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.stretch,
-                        children: [
-                          Container(
-                            color: theme.dividerColor.withOpacity(0.1),
-                            padding: const EdgeInsets.symmetric(vertical: 32.0),
-                            child: Text(
-                              'Hmmm. It seems like you\'ve reached the bottom.',
-                              textScaleFactor: state.contentFontSizeScale.textScaleFactor,
-                              textAlign: TextAlign.center,
-                              style: theme.textTheme.titleSmall,
-                            ),
+                  ),
+                if (index != widget.comments.length + 1)
+                  CommentCard(
+                    now: widget.now,
+                    selectCommentId: widget.selectedCommentId,
+                    selectedCommentPath: widget.selectedCommentPath,
+                    commentViewTree: widget.comments[index - 1],
+                    collapsedCommentSet: collapsedCommentSet,
+                    collapsed: collapsedCommentSet.contains(widget
+                            .comments[index - 1].commentView!.comment.id) ||
+                        widget.level == 2,
+                    onSaveAction: (int commentId, bool save) =>
+                        widget.onSaveAction(commentId, save),
+                    onVoteAction: (int commentId, VoteType voteType) =>
+                        widget.onVoteAction(commentId, voteType),
+                    onCollapseCommentChange: (int commentId, bool collapsed) =>
+                        onCollapseCommentChange(commentId, collapsed),
+                    onDeleteAction: (int commentId, bool deleted) =>
+                        widget.onDeleteAction(commentId, deleted),
+                  ),
+                if (index == widget.comments.length + 1) ...[
+                  if (widget.hasReachedCommentEnd == true) ...[
+                    Column(
+                      crossAxisAlignment: CrossAxisAlignment.stretch,
+                      children: [
+                        Container(
+                          color: theme.dividerColor.withOpacity(0.1),
+                          padding: const EdgeInsets.symmetric(vertical: 32.0),
+                          child: Text(
+                            widget.comments.isEmpty
+                                ? 'Oh. There are no comments.'
+                                : 'Hmmm. It seems like you\'ve reached the bottom.',
+                            textScaleFactor:
+                                state.contentFontSizeScale.textScaleFactor,
+                            textAlign: TextAlign.center,
+                            style: theme.textTheme.titleSmall,
                           ),
-                        ],
-                      )
-                    ] else ...[
-                      Column(
-                        children: [
-                          Container(
-                            padding: const EdgeInsets.symmetric(vertical: 24.0),
-                            child: const CircularProgressIndicator(),
-                          ),
-                        ],
-                      )
-                    ]
+                        ),
+                        const SizedBox(
+                          height: 160,
+                        )
+                      ],
+                    )
+                  ] else ...[
+                    Column(
+                      children: [
+                        Container(
+                          padding: const EdgeInsets.symmetric(vertical: 24.0),
+                          child: const CircularProgressIndicator(),
+                        ),
+                      ],
+                    )
                   ]
-                ]));
-          }
-        });
+                ]
+              ],
+            ),
+          );
+        }
+      },
+    );
   }
 
   int getCommentsListLength() {
@@ -185,7 +216,9 @@ class _CommentSubviewState extends State<CommentSubview> with SingleTickerProvid
       return 2; // Show post and loading indicator since no comments have been fetched yet
     }
 
-    return widget.postViewMedia != null ? widget.comments.length + 2 : widget.comments.length + 1;
+    return widget.postViewMedia != null
+        ? widget.comments.length + 2
+        : widget.comments.length + 1;
   }
 
   void onCollapseCommentChange(int commentId, bool collapsed) {
